Fixed: Assertion `M_lbase + start == ibp.first->real_start()' failed.
This is caused by (most likely a shared library) having two
symbols (functions) that have an overlapping range but different
starting points. Apparently this is possible and legit.

The current solution that at least demands that the end of the range
is the same - I suspect that even that is possible to differ too,
but lets wait till something runs into a legit case first.